backend/requirements-dev.txt
-r requirements.txt
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==5.0.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
boto3-stubs[dynamodb,s3,lambda,cognito-idp]==1.34.34
